Add DSA Stull

+++ Data Structures & Algorithms/Insertion Sort/main.cpp
@@ -4,7 +4,7 @@
 void insertion_sort(std::vector<int>& nums)
 {
     int temp, i, j;
-    for (i = 1; i < nums.size(); i++)
+    for (i = 1; i < (int)nums.size(); i++)
     {
         j = i;
         while (j > 0 && nums[j] < nums[j - 1])
@@ -19,10 +19,10 @@ void insertion_sort(std::vector<int>& nums)
 void display_vector(const std::vector<int>& vec)
 {
     std::cout << "{";
-    for(int i = 0; i < vec.size(); i++)
+    for(int i = 0; i < (int)vec.size(); i++)
     {
         std::cout << vec[i];
-        if (i < vec.size() - 1)
+        if (i < (int)vec.size() - 1)
             std::cout << ", ";
     }
     std::cout << "}" << "\n";
